==> Scripts/1. IC Data import and descriptive analysis.ipynb <==
Table of Contents:
1. Importing Libraries and Dataframe
2. Descriptive Analysis 

# 01. Importing Libraries

import pandas as pd
import numpy as np
import os

# 02. Importing Dataframe

path = r'C:\Users\franz\Documents\Exercise_4'
vars_list = 'order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order'
df = pd.read_csv(os.path.join(path, '02_Data', 'original_data', 'orders.csv'), usecols = vars_list)
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'original_data', 'products.csv'), usecols = all)

# 03. Descriptive Analysis

df.head()

df_prods.head(20)

df_prods.tail(35)

df_prods.columns

df_prods.shape

df_prods.describe()

df_prods.dtypes

### Instacart has 134 aisles in total and the data type of the department_id is integer (int64).
