==> Scripts/2. IC Data Wrangling & Subsetting.ipynb <==
Table of Contents:
1. Importing Libraries and Dataframes
2. Data Wrangling
3. Creating a Data Dictionary
4. Subsetting (create a smaller dataset from a large one by filtering)
5. Further Data Wrangling
6. Find another identifier variable that does not need to be included as a numeric variable and changing it to a suitable format
7. Replacing variables with an unintuitive name and change it without overwriting the data frame
8. Figuring out the busiest hour for placing orders
9. Determining the meaning behind values in the "department_id" using a data dictionary
10. Creating a subset containing only breakfast item sales
11. Details about customers who might be throwing dinner parties
12. Keeping track of total counts in dataframes
13. Extracting information about customer with a "user_id" of “1.”
14. Basic stats about users
15. Data Export


# 01. Importing Libraries and Dataframes

# Importing Libraries
import pandas as pd
import numpy as np
import os

# Importing Dataframes
path = r'C:\Users\franz\Documents\Exercise_4'
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'original_data', 'orders.csv'), usecols = all)
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'original_data', 'products.csv'), usecols = all)
df_dep = pd.read_csv(os.path.join(path, '02_Data', 'original_data', 'departments.csv'), usecols = all)
df_cust = pd.read_csv(os.path.join(path, '02_Data', 'original_data', 'customers.csv'), usecols = all)

# 02. Data Wrangling

# Deleting Column eval_set in orders.csv
df_ords = df_ords.drop(columns = ['eval_set'])

# Count the frequency of the days_since_prior_order columns values
df_ords['days_since_prior_order'].value_counts(dropna = False)
# There are 206.209 emtpy (NaN) cells

# renaming column order_dow in orders.csv
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_dep.head()

# transposing the df_dep 
df_dep_t = df_dep.T

df_dep_t.head()

# adding indexes to the df
df_dep_t.reset_index()

# assign the contents of row 0 to the variable new_header
new_header = df_dep_t.iloc[0]

new_header

# copy the overall df starting from index 1 (without the header)
df_dep_t_new = df_dep_t[1:]

df_dep_t_new.head()

# assign the new_header content as column names
df_dep_t_new.columns = new_header

df_dep_t_new.head()

# 03. Creating a Data Dictionary

data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

print(data_dict.get('19'))


# 04. Subsetting (create a smaller dataset from a large one by filtering)

# shows bool values for all products whether y/n they have department_id 19
df_prods['department_id']==19

# Enclosing df_prods['department_id']==19 inside df_prods[] tells Python to map the true/false procedure you witnessed 
#in the previous output onto the columns within the df_prods dataframe—the result being a list of only those values within 
#df_prods that are true
df_prods[df_prods['department_id']==19]

df_snacks = df_prods[df_prods['department_id']==19]

df_snacks.head()


#There’s a second way of achieving these same results that you might find a bit simpler. This involves use of the loc 
#function. The loc function is similar to the iloc function you used previously, but while iloc looks for an integer 
#position, loc looks for a column name. In your case, you want this function to look for instances where the value of 
#"department_id" is 19 within the df_prods database

df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]
df_snacks_2.head()

# filter for more than one id using isIn
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19, 18, 17])]
df_snacks_3.head()

# 05. Further Data Wrangling 

### 6. Find another identifier variable that does not need to be included as a numeric variable and changing it to a suitable format

#verify which data types each column has
df_ords.dtypes

# changing the data types of user_id and order_id to strings
df_ords['user_id'] = df_ords['user_id'].astype('str')
df_ords['order_id'] = df_ords['order_id'].astype('str')

# check the udpated data types
df_ords.dtypes

# Now, the describe function does not evaluate the user_id and order_id anymore, this prevents unnecessary calculations
df_ords.describe()

### 7. Replacing variables with an unintuitive name and change it without overwriting the data frame.

#renaming the column
df_ords.rename(columns = {'eval_set' : 'evaluation_set'}, inplace = True)

#checking the new column name
df_ords.head()

### 8. Figuring out the busiest hour for placing orders

# count the frequency for the order_hour_of_day variable. 
df_ords['order_hour_of_day'].value_counts(dropna = False)
# The busiest hour is 10am.

### 9. Determining the meaning behind values in the "department_id" using a data dictionary

# taking a look at all the columns of df_prods
df_prods.head()

# using my data dictionary of point 3 for df_dep to determine the meaning of department_id 4
print(data_dict.get('4'))

### 10. Creating a subset containing only breakfast item sales

# using my data dictionary to find the id of breakfast items
data_dict

df_breakfast = df_prods[df_prods['department_id']==14] #department_id 14 stands for breakfast department

# print a list of all breakfast products available for sale
df_breakfast

### 11. Details about customers who might be throwing dinner parties.

df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5, 20, 7, 12])]

df_dinner_parties

### 12. Keeping track of total counts in dataframes

df_dinner_parties.shape
# the df_dinner_parties dataframe has 7.650 rows, meaning 7.650 items of the dinner/parties category

### 13. Extracting information about customer with a "user_id" of “1.”

# extracting all the information available in df_cust for user_id = 1
df_cust_user_id_1 = df_cust[df_cust['user_id']==1]

df_cust_user_id_1

# extracting all the orders in df_ords for user_id = 1
df_ords_user_id_1 = df_ords[df_ords['user_id']=='1']

df_ords_user_id_1 

### 14. Basic stats about users

df_ords_user_id_1.describe()

### 15. Data Export

df_ords.to_csv(os.path.join(path, '02_Data','prepared_data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02_Data','prepared_data', 'departments_wrangled.csv'))
