==> Scripts/3. IC Data Consistency Checks.ipynb <==
Table of Contents: 
1. Importing Libraries and Dataframes
2. Find and fix mixed Data Types
3. Find and fix missing values
4. Find and fix duplicates
5. Exporting the clean dataset
6. Using the describe() function to verify if anything about the data looks off or should be investigated further
7. Check for mixed-type data in df_ords dataframe
8. Fixing  mixed-type data
9. Run a check for missing values in df_ords dataframe
10. Addressing the missing values
11. Run a check for duplicate values in df_ords data
12. Addressing the duplicates
13. Data Export


# 01. Importing Libraries and Dataframes


# Importing Libraries
import pandas as pd
import numpy as np
import os

# Importing Dataframes
path = r'C:\Users\franz\Documents\Exercise_4'
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'prepared_data', 'orders_wrangled.csv'), usecols = all)
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'original_data', 'products.csv'), usecols = all)

# 02. Find and fix mixed Data Types

# create a dataframe (because ours have no mixed data types)
df_test = pd.DataFrame()
# create a mixed type of column
df_test['mix'] = ['a', 'b', 1, True]

df_test

# function for checking whether a dataframe contains any mixed-type columns
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# fix the mixed data types in the mix column:
df_test['mix'] = df_test['mix'].astype('str')

# 03. Find and fix missing values

# find columns containing missing values
df_prods.isnull().sum()

# create a table with the missing datasets
df_nan = df_prods[df_prods['product_name'].isnull() == True]
df_nan

# verify the lines of rows of df_prods
df_prods.shape

# create a new dataframe without the missing rows
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# alternative code leading to the same result
#df_prods.dropna(inplace = True)

#new dataframe should have exactly 16 less rows than the original dataframe (the same as the number of missing values). Perfect!
df_prods_clean.shape

# 04. Find and fix duplicates

# find full duplicates (meaning the whole row is doubled) using the duplicated() function
df_prod_dups = df_prods_clean[df_prods_clean.duplicated()]
df_prod_dups

#checking the amount of rows before dropping the duplicates
df_prods_clean.shape

# dropping the duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# reduced by 5 lines - perfect!
df_prods_clean_no_dups.shape

## 5. Exporting the clean dataset

df_prods_clean_no_dups.to_csv(os.path.join(path, '02_Data','prepared_data', 'products_checked_for_consistency.csv'))

## 6. Using the describe() function to verify if anything about the data looks off or should be investigated further

df_ords.describe()

#### orders_day_of_week: can not have a maximum value of > 6 and minimum value of < 0. This is not the case here. Double checking the 50th percentile, also no strange value can be found since this value is supposed to be 3. 
#### the days_since_prior_order can not contain values smaller than 0 (negative values). This would imply a negative number of days. This is also not the case. 
#### the order_hour_of_day must have a minimum value of 0 and a maximum value of 24, which is given in our df_orders. 

# 7. Check for mixed-type data in df_ords dataframe

# function for checking whether a df_ords contains any mixed-type columns
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)


#### No mixed-type data could be found in df_ords.

# 8. Fixing  mixed-type data

#### As no mixed-type data could be found, this task is obsolete. 

# 9. Run a check for missing values in df_ords dataframe

# find missing values
df_ords.isnull().sum()

# create a table with the missing datasets
df_nan_ords = df_ords[df_ords['days_since_prior_order'].isnull() == True]
df_nan_ords.head()

#### The missing values were found in the days_since_prior_order column. Those are orders without a prior order. 

# 10. Addressing the missing values

#### They will not be fixed as those are supposed to be customers who have only placed one or less orders so far. Therefore, it makes sense that those cells are empty.  

# 11. Run a check for duplicate values in df_ords data

# checking the duplicates
df_dups_ords = df_ords[df_ords.duplicated()]
df_dups_ords

#### No full duplicate rows were found.

# 12. Addressing the duplicates

#### As no duplicates were found, this is not necessary.

# 13. Data Export

#### I haven´t changed anything in df_ords, so I kept my orders_wrangled.csv from exercise 4.4 but renamed it as: orders_wrangled_and_checked_for_consistency. The df_prods_clean_no_dups was exported previously (see step 5. Exporting the clean dataset)
