==> Scripts/4. IC Combining and Exporting Data - Task.ipynb <==
Table of Contents:
1. Importing Libraries and Dataframes
2. Preparations for Merge
3. Performing the Merge

# 01. Importing Libraries and Dataframes

# Importing Libraries
import pandas as pd
import numpy as np
import os

# Importing Dataframes
path = r'C:\Users\franz\Documents\Exercise_4'
df_orders_products_combined = pd.read_pickle(os.path.join(path, '02_Data', 'prepared_data', 'orders_products_combined.pkl'))
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'prepared_data', 'products_checked_for_consistency.csv'), usecols = all)

# 02. Preparations for Merge

# Check the shape of the imported dataframe (it should be the same as the one you exported (32434489, 1)
df_orders_products_combined.shape

#check the column names of df_prods to determine a suitable method for the merge
df_prods.head()

#check the column names of df_orders_products_combined to determine a suitable method for the merge
df_orders_products_combined.head()

df_orders_products_combined.info()

# change data types to free some memory for the merge

df_orders_products_combined['product_id'] =df_orders_products_combined['product_id'].astype('int32')
df_orders_products_combined['reordered']=df_orders_products_combined['reordered'].astype('int8')
df_orders_products_combined['add_to_cart_order']=df_orders_products_combined['add_to_cart_order'].astype('int32')
df_orders_products_combined['order_id']=df_orders_products_combined['order_id'].astype('int32')
df_orders_products_combined['user_id'] =df_orders_products_combined['user_id'].astype('int32')
df_orders_products_combined['order_number'] =df_orders_products_combined['order_number'].astype('int32')
df_orders_products_combined['orders_day_of_week'] =df_orders_products_combined['orders_day_of_week'].astype('int32')
df_orders_products_combined['order_hour_of_day'] =df_orders_products_combined['order_hour_of_day'].astype('int32')


df_prods['product_id'] =df_prods['product_id'].astype('int32')
df_prods['aisle_id'] = df_prods['aisle_id'].astype('int8')
df_prods['department_id'] = df_prods['department_id'].astype('int8')
df_prods['prices'] = df_prods['prices'].astype('float16')

df_orders_products_combined.info()

# 03. Performing the Merge

# combine the df_orders_products_combined with the df_prods using a default inner merge, using an 'exists' string instead 
# of boolean True for the indicator because _merge column already exists in previous dataframe

prods_ords_merged = df_prods.merge(df_orders_products_combined, on = ['product_id'], indicator='exists')

# Confirm the results of the merge using the merge flag (check if 'exists' column is available)
prods_ords_merged.head()

prods_ords_merged.info()

# export into pickle format due to large file size of 2.3+GB

prods_ords_merged.to_pickle(os.path.join(path, '02_data','prepared_data', 'orders_products_merged.pkl'))
