==> Scripts/5. IC Deriving New Variables.ipynb <==
Table of Contents: 
1. Importing Libraries and Dataframes
2. if-statements and loops
- If-Statement with user-defined functions
- If-Statement with loc() Function
- If statements with for-Loops
3. Creating the “price_label” and “busiest_day” columns
4. “Busiest day” to become “Busiest days” (plural) 
5. Check the values of this new column for accuracy
6. Creating a new column called “busiest_period_of_day.”
7. Printing the frequency for this new column and data export

# 01. Importing Libraries and Dataframes

# Importing Libraries
import pandas as pd
import numpy as np
import os
# disable a new warning with the following assignment
pd.options.mode.chained_assignment = None  # default='warn'

# Importing Dataframes
path = r'C:\Users\franz\Documents\Exercise_4'
df_ords_prods = pd.read_pickle(os.path.join(path, '02_Data', 'prepared_data', 'orders_products_merged.pkl'))

 # 02. if-statements and loops

## 2.1 If-Statement with user-defined functions

# sing user-defined functions on a large dataframe can lead to memory issues or trouble with processing power. 
# And your Instacart dataframe is, indeed, large. To avoid any potential issues, let’s just work with a subset of the dataframe for now—the first one million rows.
df = df_ords_prods[:1000000]

# simple if statement 

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'
    
    
#If the value within the “prices” column within the given row is less than or equal to 5, then return the string “Low-range product.”
#Or else, if the value within the “prices” column within the given row is greater than 5 and less than or equal to 15, then return the string “Mid-range product.”
#Or else, if the value of the “prices” column within the given row is greater than 15, then return the string “High-range product.”
#Or else, return the string “Not enough data.”

df['price_label'] = df.apply(price_label, axis=1)

df.head()

df['price_label'].value_counts(dropna = False)

df['prices'].max()

crosstab_price_label = pd.crosstab(df['product_name'], df['price_label'], dropna = False)
crosstab_price_label

# copy the table straight to your clipboard, allowing you to then paste it into Excel for easy viewing. 
crosstab_price_label.to_clipboard()

## 2.2 If-Statement with loc() Function

### Applying the loc() function to a part of the dataframe

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

### Applying the loc() function to the complete dataframe

df_ords_prods.loc[df_ords_prods['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods.loc[(df_ords_prods['prices'] <= 15) & (df_ords_prods['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df_ords_prods.loc[df_ords_prods['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods['price_range_loc'].value_counts(dropna = False)

## 2.3 If statements with for-Loops

#need to know on which day most orders take place. You can find this out by printing the frequency of the “orders_day_of_the_week” column, which would look like this:
df_ords_prods['orders_day_of_week'].value_counts(dropna = False)

#Now, you have some numbers. But what do these numbers mean? In your project brief, you can see that the value 0 means Saturday. This value has the highest frequency, which means that Saturday is the busiest day. Meanwhile, the 4 value has the lowest frequency. A value of 4, here, refers to Wednesday, meaning that Wednesday is the slowest day when it comes to Instacart app orders.
#You want to use this information to create a new column, “busiest day,” that will contain one of three different values: “Busiest day,” “Least busy,” and “Regularly busy.” This can be done by way of a for-loop. The loop will run through every row in the “orders_day_of_the_week” column, compare the value in it with what you know are the busiest and slowest days, and assign it the corresponding string value.

result = []

for value in df_ords_prods["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

# create a new column and assign the result - list
df_ords_prods['busiest_day'] = result

# check how often each value appears
df_ords_prods['busiest_day'].value_counts(dropna = False)

#  print the frequency of this new column and cross-check it with the frequency of “orders_day_of_the_week” you printed before
df_ords_prods['orders_day_of_week'].value_counts(dropna = False)

## 3. Creating the “price_label” and “busiest_day” columns

df_ords_prods.head()

## 4. “Busiest day” to become “Busiest days” (plural) 

This label should correspond with the two busiest days of the week as opposed to the single busiest day. At the same time, they’d also like to know the two slowest days. Create a new column for this using a suitable method.


# create a list of:
# Busiest days: 0 and 1 
# Least busy days: 3 and 4
# Regularly busy: 2, 5, 6

result_task2 = []

for value in df_ords_prods["orders_day_of_week"]:
  if value == 0 or value == 1:
    result_task2.append("Busiest days")
  elif value == 4 or value == 3:
    result_task2.append("Slowest days")
  else:
    result_task2.append("Regularly busy")


# create a new column and assign the result - list
df_ords_prods['busiest_days'] = result_task2

## 5. Check the values of this new column for accuracy

# check how often each value appears in new column busiest_days
df_ords_prods['busiest_days'].value_counts(dropna = False)

# check how often the values appear in base-column 
df_ords_prods['orders_day_of_week'].value_counts(dropna = False)

#### busiest_days: 32.404.859 values 
#### orders_day_of_week: 32.404.859 values 
#### As both columns have the same amount of values, no issues can be found.

## 6. Creating a new column called “busiest_period_of_day.”

When too many users make Instacart orders at the same time, the app freezes. The senior technical officer at Instacart wants you to identify the busiest hours of the day. Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” Create a new column containing these labels called “busiest_period_of_day.”

#need to know on which day most orders take place. You can find this out by printing the frequency of the “order_hour_of_day” column, which would look like this:
df_ords_prods['order_hour_of_day'].value_counts(dropna = False)

# Most orders: 10 
# Fewest orders: 3


result_task3 = []

for value in df_ords_prods["order_hour_of_day"]:
  if value == 10:
    result_task3.append("Most orders")
  elif value == 3:
    result_task3.append("Fewest orders")
  else:
    result_task3.append("Average orders")


# create a new column and assign the result - list
df_ords_prods['busiest_period_of_day'] = result_task3

## 7. Printing the frequency for this new column and data export

df_ords_prods['busiest_period_of_day'].value_counts(dropna = False)

# Data Export: 
df_ords_prods.to_pickle(os.path.join(path, '02_data','prepared_data', 'orders_products_incl_new_variables.pkl'))
