==> Scripts/7. IC Datavisualization - Part 1.ipynb <==
Table of Contents: 
1. Importing Libraries and Dataframes
2. Wrangling the data so that it follows consistent logic
3. Completing the fundamental data quality and consistency checks
4. Combining the customer data with the rest of prepared Instacart data


### 1. Importing Libraries and Dataframes

# Importing Analysis Libraries
import pandas as pd
import numpy as np
import os

# Importing Dataframes
path = r'C:\Users\franz\Documents\Exercise_4'
df = pd.read_csv(os.path.join(path, '02_Data', 'original_data', 'customers_4.9.csv'), usecols = all)
ords_prods_merge_old = pd.read_pickle(os.path.join(path, '02_Data', 'prepared_data', 'orders_products_incl_new_variables_4.9_dropped.pkl'))

### 2. Wrangling the data so that it follows consistent logic

for example, rename columns with illogical names and drop columns that don’t add anything to your analysis.

df.head()

# renaming columns with illogical names
df.rename(columns = {'First Name':'first_name', 'Surnam':'last_name', 'Gender':'gender', 'State':'state', 'Age':'age', 'STATE':'state', 'fam_status':'marital_status'}, inplace = True)

#### drop columns that don´t add anything to your analysis: I did not find any irrelevant column

df.head()

### 3. Completing the fundamental data quality and consistency checks

for example, check for and address missing values and duplicates, and convert any mixed-type data.

#### Mixed data types:

# Checking for mixed-type columns

for col in df.columns.tolist():
  weird = (df[[col]].applymap(type) != df[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df[weird]) > 0:
    print (col)

# Fixing the mixed-type column "first_name" and assigning the datatype string, as names are always strings

df['first_name'] = df['first_name'].astype('str')

# Checking for mixed-type columns again

for col in df.columns.tolist():
  weird = (df[[col]].applymap(type) != df[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df[weird]) > 0:
    print (col)

df.dtypes

#### Now, there are no mixed data types anymore.

# Find missing values:
df.isnull().sum()

#### We do not have any missing values

# looking for full duplicates within the dataframe:
df_dups = df[df.duplicated()]

df_dups.head()

#### We do not have any duplicate rows

# data export
df.to_pickle(os.path.join(path, '02_data','prepared_data', 'customers.pkl'))

### 4. Combining the customer data with the rest of prepared Instacart data

#### Find the common column(s):

ords_prods_merge_old.columns

df.columns

#### The key column is user_id

#### Performing some pre-checks:

df.dtypes

ords_prods_merge_old.dtypes

#assigning int32 type to the user_id column in df (as it is the case in ords_prods_merge_old)
#change data types to free some memory for the merge
df['user_id'] =df['user_id'].astype('int32')
df['age'] =df['age'].astype('int32')
df['income'] =df['income'].astype('int32')

df.head()

df.dtypes

df.shape

ords_prods_merge_old.shape

Deleting some columns in ords_prods_merge_old to save memory

ords_prods_merge_old = ords_prods_merge_old.drop(columns = ['_merge'])

ords_prods_merge_old = ords_prods_merge_old.drop(columns = ['Unnamed: 0_x'])

ords_prods_merge_old = ords_prods_merge_old.drop(columns = ['Unnamed: 0_y'])

ords_prods_merge_old = ords_prods_merge_old.drop(columns = ['exists'])

ords_prods_merge_old['add_to_cart_order'] =ords_prods_merge_old['add_to_cart_order'].astype('int16')
ords_prods_merge_old['days_since_prior_order'] =ords_prods_merge_old['days_since_prior_order'].astype('float16')
ords_prods_merge_old['median_days_since_prior_order'] =ords_prods_merge_old['median_days_since_prior_order'].astype('float16')

# Data Export: 
# ords_prods_merge_old.to_pickle(os.path.join(path, '02_data','prepared_data', 'orders_products_incl_new_variables_4.9_dropped.pkl'))

### Performing the actual merge:

ords_prods_merge_old.info()

# direct merge does not work due to memory errors, thus it will not be applied
ords_prods_cust_merge = df.merge(ords_prods_merge_old, on = 'user_id', indicator = 'exists_4.9')
